feat: complete M13 integration, docs, benchmarks and release milestone
- E2E integration tests: DLQ lifecycle, workflow timeout/retry, backpressure, secret rotation
- Performance benchmarks: Event Bus 372K/s, Pipeline 3.5us, WAL batch 57K/s, etc.
- Developer documentation: README, API, CLI, CONFIG, SKILL, PLUGIN, WORKFLOW, ARCHITECTURE
- SDK crate with prelude re-exports and hello-skill example
- CI/CD (GitHub Actions: clippy + test + doc on Linux/macOS)
- CHANGELOG.md with SemVer policy
- All path dependencies versioned for cargo publish
- Clippy fixes across all crates (Rust 1.88 idioms, derive Default, if-let chains)
- Milestone progress updated to 100%

+++ crates/runtime/src/metrics.rs
@@ -128,11 +128,11 @@ impl MetricsRegistry {
             .with_label_values(&["low"])
             .set(bus.queue_depth.low as i64);
         self.throughput.reset();
-        self.throughput.inc_by(bus.throughput as u64);
+        self.throughput.inc_by(bus.throughput);
         self.discarded.reset();
-        self.discarded.inc_by(bus.discarded_count as u64);
+        self.discarded.inc_by(bus.discarded_count);
         self.duplicate.reset();
-        self.duplicate.inc_by(bus.duplicate_count as u64);
+        self.duplicate.inc_by(bus.duplicate_count);
         self.subscription_count.set(bus.subscription_count as i64);
         self.retry_queue_depth.set(bus.retry_queue_depth as i64);
         self.dlq_depth.set(dlq_depth_val as i64);
